Add executables for DOS/Windows 10
Goterm.exe and goproxyterm.exe we not present for DOS / Windows 10
Without these, PF6 is difficult to use.
Getting Go language to compile these properly was difficult due to differences in Linus and Windows10, and the lack of detailed instructions for Win10.  I believe the Win10 instructions for GOPATH, GOROOT and MYGO need refinement.